Use WorkerInvokeCallInRib in component constraints
