+++ golem-rib/src/compiler/mod.rs
@@ -16,17 +16,18 @@ pub use byte_code::*;
 use golem_wasm_ast::analysis::AnalysedExport;
 pub use ir::*;
 pub use type_with_unit::*;
+pub use compiler_output::*;
+pub use worker_invoke_calls::*;
 
-use crate::compiler::function_calls::WorkerInvokeCallsInRib;
 use crate::type_registry::FunctionTypeRegistry;
 use crate::{Expr, InferredExpr, RibInputTypeInfo};
-use golem_api_grpc::proto::golem::rib::CompilerOutput as ProtoCompilerOutput;
 
 mod byte_code;
 mod desugar;
-mod function_calls;
+mod worker_invoke_calls;
 mod ir;
 mod type_with_unit;
+mod compiler_output;
 
 pub fn compile(
     expr: &Expr,
@@ -76,49 +77,4 @@ pub fn compile_with_limited_globals(
         byte_code,
         global_input_type_info,
     })
-}
-
-#[derive(Debug, Clone)]
-pub struct CompilerOutput {
-    pub worker_invoke_calls: Option<WorkerInvokeCallsInRib>,
-    pub byte_code: RibByteCode,
-    pub global_input_type_info: RibInputTypeInfo,
-}
-
-impl TryFrom<ProtoCompilerOutput> for CompilerOutput {
-    type Error = String;
-
-    fn try_from(value: ProtoCompilerOutput) -> Result<Self, Self::Error> {
-        let proto_rib_input = value.rib_input.ok_or("Missing rib_input")?;
-        let proto_byte_code = value.byte_code.ok_or("Missing byte_code")?;
-        let rib_input = RibInputTypeInfo::try_from(proto_rib_input)?;
-        let byte_code = RibByteCode::try_from(proto_byte_code)?;
-        let worker_invoke_calls = if let Some(value) = value.worker_invoke_calls {
-            Some(WorkerInvokeCallsInRib::try_from(value)?)
-        } else {
-            None
-        };
-
-        Ok(CompilerOutput {
-            worker_invoke_calls,
-            byte_code,
-            global_input_type_info: rib_input,
-        })
-    }
-}
-
-impl From<CompilerOutput> for ProtoCompilerOutput {
-    fn from(value: CompilerOutput) -> Self {
-        ProtoCompilerOutput {
-            byte_code: Some(golem_api_grpc::proto::golem::rib::RibByteCode::from(
-                value.byte_code,
-            )),
-            rib_input: Some(golem_api_grpc::proto::golem::rib::RibInputType::from(
-                value.global_input_type_info,
-            )),
-            worker_invoke_calls: value
-                .worker_invoke_calls
-                .map(golem_api_grpc::proto::golem::rib::WorkerInvokeCallsInRib::from),
-        }
-    }
 }
