+++ golem-rib/src/type_registry.rs
@@ -18,6 +18,7 @@ use golem_api_grpc::proto::golem::rib::registry_key::KeyType;
 use golem_wasm_ast::analysis::AnalysedType;
 use golem_wasm_ast::analysis::{AnalysedExport, TypeVariant};
 use std::collections::{HashMap, HashSet};
+use serde::{Deserialize, Serialize};
 
 // A type-registry is a mapping from a function name (global or part of an interface in WIT)
 // to the registry value that represents the type of the name.
@@ -171,7 +172,7 @@ impl FunctionTypeRegistry {
     }
 }
 
-#[derive(Hash, Eq, PartialEq, Clone, Debug)]
+#[derive(Hash, Eq, PartialEq, Clone, Debug, Serialize, Deserialize)]
 pub enum RegistryKey {
     FunctionName(String),
     FunctionNameWithInterface {
